refactor(capabilities): :recycle: migrate command-palette cluster to component-store pattern
CommandPalette: ComponentStore<{open,isSearching,selectedValue,search,
easterEggLines},{close,stopPropagation,handleSearchInput,handleOpenChat,
handleSearchResultSelect,setSelectedValue}> composing useSearch, useRouter.
Preserves searchEasterEgg + SearchEasterEggComponent inversion props — no
features/ import. ToggleMotionItem and CustomizeMatrixRainItem consumer-count
nesting: private sub-folders inside command-palette/command-palette/.
SearchResultItem sub-component eliminates per-item arrow functions in JSX.
Updated command-palette/index.ts to re-export from new sub-folder.
