+++ src/components/design-system/organism/command-palette/command-palette/customize-matrix-rain-item/customize-matrix-rain-item.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import { Command } from "cmdk";
+import { TerminalLine } from "@/components/design-system/atoms/typography/terminal-blocks";
+import { MdTune } from "react-icons/md";
+import { useCustomizeMatrixRainItemStore } from "./use-customize-matrix-rain-item-store";
+
+const ITEM_CLASS =
+    "px-4 py-2 cursor-pointer aria-selected:bg-accent-alpha-10 aria-selected:border-l-2 aria-selected:border-accent transition-colors duration-100";
+
+interface CustomizeMatrixRainItemProps {
+    onClose: () => void;
+}
+
+export const CustomizeMatrixRainItem = ({ onClose }: CustomizeMatrixRainItemProps) => {
+    const { state, effects } = useCustomizeMatrixRainItemStore(onClose);
+    const { visible } = state;
+    const { handleSelect } = effects;
+
+    if (!visible) {
+        return null;
+    }
+
+    return (
+        <Command.Item value="customize matrix rain" className={ITEM_CLASS} onSelect={handleSelect}>
+            <TerminalLine>
+                <MdTune className="inline mr-2 mb-0.5" />
+                {">"} Customize Matrix Rain
+            </TerminalLine>
+        </Command.Item>
+    );
+};
