+++ src/components/design-system/organism/command-palette/command-palette/customize-matrix-rain-item/index.ts
@@ -1,0 +1,1 @@
+export { CustomizeMatrixRainItem } from "./customize-matrix-rain-item";
