+++ src/components/design-system/organism/command-palette/command-palette/search-result-item/index.ts
@@ -1,0 +1,1 @@
+export { SearchResultItem } from "./search-result-item";
