+++ src/components/design-system/organism/command-palette/command-palette/search-result-item/use-search-result-item-store.ts
@@ -1,0 +1,15 @@
+"use client";
+
+import { EffectsStore } from "@/types/component-store";
+
+interface SearchResultItemEffects {
+    handleSelect: () => void;
+}
+
+export const useSearchResultItemStore = (slug: string, onSelect: (slug: string) => void): EffectsStore<SearchResultItemEffects> => {
+    const handleSelect = () => onSelect(slug);
+
+    return {
+        effects: { handleSelect },
+    };
+};
