+++ src/components/design-system/organism/command-palette/command-palette/toggle-motion-item/index.ts
@@ -1,0 +1,1 @@
+export { ToggleMotionItem } from "./toggle-motion-item";
